refactor(site-header-core): specs file, test and story adapted - TWIG-151
