+++ src/ec/packages/ec-component-site-header-core/adapter.js
@@ -1,14 +1,10 @@
 /* eslint-disable import/no-extraneous-dependencies, no-param-reassign */
-import specEnglishData from '@ecl/ec-specs-site-header-core/demo/data--en';
-import specFrenchData from '@ecl/ec-specs-site-header-core/demo/data--fr';
-
-const defaultSprite = '/icons.svg';
-
-const englishBanner = '/logo--en.svg';
-const frenchBanner = '/logo--fr.svg';
-
 const adapter = initialData => {
   const adaptedData = JSON.parse(JSON.stringify(initialData));
+
+  const defaultSprite = '/icons.svg';
+  const englishBanner = '/logo--en.svg';
+  const frenchBanner = '/logo--fr.svg';
 
   if (adaptedData.loginToggle) {
     adaptedData.login_toggle = {
@@ -37,7 +33,7 @@ const adapter = initialData => {
     item.path = item.href;
     delete item.href;
     if (item.isActive) {
-      item.active = true; // eslint-disable-line no-param-reassign
+      item.active = true;
       delete item.isActive;
     }
   });
@@ -62,5 +58,4 @@ const adapter = initialData => {
   return adaptedData;
 };
 
-export const englishData = adapter(specEnglishData);
-export const frenchData = adapter(specFrenchData);
+export default adapter;
