+++ src/ec/packages/ec-component-site-header-core/demo/data--en.js
@@ -1,0 +1,4 @@
+import specs from '@ecl/ec-specs-site-header-core/demo/data--en';
+import adapter from '../adapter';
+
+export default adapter(specs);
